+++ module-services/service-eink/messages/EinkMessage.hpp
@@ -20,8 +20,8 @@ namespace seink {
  */
 class EinkMessage: public sys::DataMessage {
 public:
-	EinkMessage( MessageType messageType ) : sys::DataMessage( static_cast<uint32_t>(messageType)) {};
-	virtual ~EinkMessage() {};
+  EinkMessage(MessageType messageType) : sys::DataMessage(messageType){};
+  virtual ~EinkMessage(){};
 };
 
 } /* namespace seink */
